Fixes following feedback
